# src/main/java/com/auspost/demo/exception/ErrorResponse.java
package com.auspost.demo.exception;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ErrorResponse {
    private String code;
    private String message;
}